Delete the "new" QueryManager, which was unused, and rename "OldQueryManager" to just "QueryManager"
