+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/querymgr/LiveQueries.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.firebase.dataconnect.oldquerymgr
+package com.google.firebase.dataconnect.querymgr
 
 import com.google.firebase.dataconnect.*
 import com.google.firebase.dataconnect.core.Logger
