--- firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/querymgr/TypedActiveQueryKey.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.firebase.dataconnect.querymgr
-
-import kotlinx.serialization.DeserializationStrategy
-
-internal class TypedActiveQueryKey<Data>(val dataDeserializer: DeserializationStrategy<Data>) {
-  override fun equals(other: Any?) =
-    other is TypedActiveQueryKey<*> && other.dataDeserializer === dataDeserializer
-
-  override fun hashCode() = System.identityHashCode(dataDeserializer)
-
-  override fun toString() = "TypedActiveQueryKey(dataDeserializer=$dataDeserializer)"
-}
